+++ src/main/java/org/cojen/maker/BaseType.java
@@ -267,6 +267,11 @@ abstract class BaseType implements Type, Typed {
     }
 
     /**
+     * @return false if the class couldn't be found by loading it
+     */
+    abstract boolean classExists();
+
+    /**
      * Returns the type code used by the stack map table attribute.
      *
      * @return SM_*
@@ -586,8 +591,13 @@ abstract class BaseType implements Type, Typed {
                 }
             }
 
-            throw new IllegalStateException
-                ("No matching methods found for: " + name() + '.' + methodName);
+            String message = "No matching methods found for: " + name() + '.' + methodName;
+
+            if (!classExists()) {
+                message = message + " (class not found)";
+            }
+
+            throw new IllegalStateException(message);
         }
 
         var b = new StringBuilder()
@@ -1021,6 +1031,11 @@ abstract class BaseType implements Type, Typed {
         }
 
         @Override
+        boolean classExists() {
+            return true;
+        }
+
+        @Override
         int stackMapCode() {
             return mStackMapCode;
         }
@@ -1167,6 +1182,11 @@ abstract class BaseType implements Type, Typed {
         }
 
         @Override
+        boolean classExists() {
+            return true;
+        }
+
+        @Override
         int stackMapCode() {
             return SM_NULL;
         }
@@ -1263,6 +1283,11 @@ abstract class BaseType implements Type, Typed {
                 }
             }
             return clazz;
+        }
+
+        @Override
+        boolean classExists() {
+            return classType() != null;
         }
 
         @Override
@@ -1423,6 +1448,11 @@ abstract class BaseType implements Type, Typed {
         }
 
         @Override
+        boolean classExists() {
+            return classType() != null;
+        }
+
+        @Override
         boolean isHidden() {
             Class clazz = classType();
             return clazz != null && clazz.isHidden();
@@ -1545,7 +1575,7 @@ abstract class BaseType implements Type, Typed {
                     return field;
                 }
                 if (field.equals(existing)) {
-                    return existing;
+                    throw new IllegalStateException("Field is already defined: " + name);
                 }
             }
 
@@ -1808,7 +1838,7 @@ abstract class BaseType implements Type, Typed {
                     return method;
                 }
                 if (method.equals(existing)) {
-                    return existing;
+                    throw new IllegalStateException("Method is already defined: " + existing);
                 }
             }
 
@@ -1956,6 +1986,11 @@ abstract class BaseType implements Type, Typed {
         }
 
         @Override
+        boolean classExists() {
+            return true;
+        }
+
+        @Override
         TheClassMaker makerType() {
             return mMakerRef.get();
         }
@@ -1964,7 +1999,11 @@ abstract class BaseType implements Type, Typed {
         BaseType superType() {
             BaseType superType = mSuperType;
             if (superType == null) {
-                mSuperType = superType = makerType().superType();
+                TheClassMaker makerType = makerType();
+                BaseType candidate = makerType.superTypeNonNull();
+                if (!candidate.equals(makerType.type())) {
+                    mSuperType = superType = candidate;
+                }
             }
             return superType;
         }