+++ src/main/java/org/cojen/maker/ClassMaker.java
@@ -231,6 +231,7 @@ public interface ClassMaker extends Maker {
      * @return this
      * @throws IllegalArgumentException if the type is unsupported
      * @throws IllegalStateException if already assigned
+     * @see <a href="package-summary.html#types-and-values-heading">Types and Values</a>
      */
     ClassMaker extend(Object superClass);
 
@@ -241,6 +242,7 @@ public interface ClassMaker extends Maker {
      * @param iface non-null type, specified by a Class or a String, etc.
      * @throws IllegalArgumentException if the type is unsupported
      * @return this
+     * @see <a href="package-summary.html#types-and-values-heading">Types and Values</a>
      */
     ClassMaker implement(Object iface);
 
@@ -256,6 +258,7 @@ public interface ClassMaker extends Maker {
      * @param subclass non-null type, specified by a Class or a String, etc.
      * @return this
      * @throws IllegalArgumentException if the type is unsupported
+     * @see <a href="package-summary.html#types-and-values-heading">Types and Values</a>
      */
     ClassMaker permitSubclass(Object subclass);
 
@@ -265,6 +268,7 @@ public interface ClassMaker extends Maker {
      * @param type a class or name
      * @throws IllegalArgumentException if the type is unsupported
      * @throws IllegalStateException if field is already defined
+     * @see <a href="package-summary.html#types-and-values-heading">Types and Values</a>
      */
     FieldMaker addField(Object type, String name);
 
@@ -274,6 +278,8 @@ public interface ClassMaker extends Maker {
      * @param retType a class or name; can be null if the method returns void
      * @param paramTypes classes or names
      * @throws IllegalArgumentException if a type is unsupported
+     * @throws IllegalStateException if method is already defined
+     * @see <a href="package-summary.html#types-and-values-heading">Types and Values</a>
      */
     MethodMaker addMethod(Object retType, String name, Object... paramTypes);
 
@@ -288,6 +294,7 @@ public interface ClassMaker extends Maker {
      * Add a method to this class.
      *
      * @param type defines the return type and parameter types
+     * @throws IllegalStateException if method is already defined
      */
     default MethodMaker addMethod(String name, MethodType type) {
         return addMethod(type.returnType(), name, (Object[]) type.parameterArray());
@@ -297,6 +304,7 @@ public interface ClassMaker extends Maker {
      * Add a method to this class.
      *
      * @param desc defines the return type and parameter types
+     * @throws IllegalStateException if method is already defined
      */
     default MethodMaker addMethod(String name, MethodTypeDesc desc) {
         return addMethod(desc.returnType(), name, (Object[]) desc.parameterArray());
@@ -307,6 +315,8 @@ public interface ClassMaker extends Maker {
      *
      * @param paramTypes classes or names
      * @throws IllegalArgumentException if a type is unsupported
+     * @throws IllegalStateException if constructor is already defined
+     * @see <a href="package-summary.html#types-and-values-heading">Types and Values</a>
      */
     MethodMaker addConstructor(Object... paramTypes);
 
@@ -322,6 +332,7 @@ public interface ClassMaker extends Maker {
      *
      * @param type defines the parameter types
      * @throws IllegalArgumentException if the return type isn't void
+     * @throws IllegalStateException if constructor is already defined
      */
     default MethodMaker addConstructor(MethodType type) {
         if (type.returnType() != void.class) {
@@ -335,6 +346,7 @@ public interface ClassMaker extends Maker {
      *
      * @param desc defines the parameter types
      * @throws IllegalArgumentException if the return type isn't void
+     * @throws IllegalStateException if constructor is already defined
      */
     default MethodMaker addConstructor(MethodTypeDesc desc) {
         if (!ConstantDescs.CD_void.equals(desc.returnType())) {